preparation for state handler refactoring
