+++ src/sloby-editor-framework/tools/test_tool/index.ts
@@ -1,0 +1,12 @@
+import { SlobyBaseTool } from "../../utils/abstracts";
+
+export default class TestTool implements SlobyBaseTool {
+  constructor() {}
+
+   initializeTool(tool: string) {
+    console.log(`${tool} has been initialized`)
+    return true
+  }
+
+  status() { return "" }
+}
