+++ src/sloby-editor-framework/tools/tools_store.ts
@@ -1,5 +1,4 @@
 import TextCreationTool from "./text-creator";
-
 /**
  * Creating a store for the tools, and that's gonna be the refferr
  */
